Rename constants -> statusMarks & export statusMarks.
